Added `GIMesh::merge_vertices` `vertex_count` now returns a `u32` Added `GIMesh::get_or_add_vertex` Updated `README.md` to use `GIMesh::slice` and `GIMesh::seperate`
deprecated:
`bevy_mops::merge_meshes`
`bevy_mops::merge_vertices`
`bevy_mops::seperate`
`bevy_mops::slice`

+++ src/seperate.rs
@@ -9,12 +9,12 @@ pub fn seperate(a: &GIMesh, b: &GIMesh) -> SeperateOutput {
     let mut output = SeperateOutput {
         inside: GIMesh {
             indices: Vec::with_capacity(a.index_count()),
-            vertices: Vec::with_capacity(a.vertex_count()),
+            vertices: Vec::with_capacity(a.vertex_count() as usize),
             inverse_model: a.inverse_model,
         },
         outside: GIMesh {
             indices: Vec::with_capacity(a.index_count()),
-            vertices: Vec::with_capacity(a.vertex_count()),
+            vertices: Vec::with_capacity(a.vertex_count() as usize),
             inverse_model: a.inverse_model,
         },
     };
@@ -53,6 +53,7 @@ pub fn seperate(a: &GIMesh, b: &GIMesh) -> SeperateOutput {
             &mut output.inside
         };
 
+        // FIXME: This will readd vertices already added to `mesh`
         let a = mesh.add_vertex(a_verts[0].clone());
         let b = mesh.add_vertex(a_verts[1].clone());
         let c = mesh.add_vertex(a_verts[2].clone());
